refactor(source): services singulares + loaders + tests
